fold decorators_tests.test_solid -> test_op
